Runtime/efi: Started implementing the functionality to parse, create and modify EFI signature databases, bugref:9580 [build fix]
svn:sync-xref-src-repo-rev: r145835

+++ src/VBox/Runtime/common/efi/efisignaturedb.cpp
@@ -1,4 +1,4 @@
-/* $Id: efisignaturedb.cpp 90285 2021-07-22 11:59:15Z alexander.eichner@oracle.com $ */
+/* $Id: efisignaturedb.cpp 90287 2021-07-22 13:16:15Z alexander.eichner@oracle.com $ */
 /** @file
  * IPRT - EFI signature database helpers.
  */
@@ -501,7 +501,7 @@ RTDECL(int) RTEfiSigDbAddSignatureFromBuf(RTEFISIGDB hEfiSigDb, RTEFISIGTYPE enm
     if (   !pDesc->cbSig
         || pDesc->cbSig - sizeof(EFI_SIGNATURE_DATA) == cbBuf)
     {
-        PRTEFISIGNATURE pSig = rtEfiSigDbAllocSignature(pUuidOwner, cbBuf);
+        PRTEFISIGNATURE pSig = rtEfiSigDbAllocSignature(pUuidOwner, (uint32_t)cbBuf);
         if (pSig)
         {
             memcpy(&pSig->abSignature[0], pvBuf, cbBuf);
